refactor: model, service, repository, DTOs and controller of Sala De Espera
