+++ healsync/src/main/java/com/pi/healsync/DTO/SalaDeEsperaResponseDTO.java
@@ -1,7 +1,12 @@
-package main.java.com.pi.healsync.DTO;
+package com.pi.healsync.DTO;
+
+import lombok.Getter;
+import lombok.Setter;
 
 import java.util.UUID;
 
+@Getter
+@Setter
 public class SalaDeEsperaResponseDTO {
 
     private UUID id;
@@ -12,22 +17,6 @@ public class SalaDeEsperaResponseDTO {
 
     public SalaDeEsperaResponseDTO(UUID id, String sala) {
         this.id = id;
-        this.sala = sala;
-    }
-
-    public UUID getId() {
-        return id;
-    }
-
-    public void setId(UUID id) {
-        this.id = id;
-    }
-
-    public String getSala() {
-        return sala;
-    }
-
-    public void setSala(String sala) {
         this.sala = sala;
     }
 }
