+++ healsync/src/main/java/com/pi/healsync/services/SalaDeEsperaService.java
@@ -1,45 +1,45 @@
-package main.java.com.pi.healsync.services;
+package com.pi.healsync.services;
 
-import java.util.Optional;
+import java.util.List;
 import java.util.UUID;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import javax.management.RuntimeErrorException;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.pi.healsync.exceptions.NoSuchException;
-import com.pi.healsync.exceptions.ObjectNotCreated;
 import com.pi.healsync.models.SalaDeEspera;
 import com.pi.healsync.repositories.SalaDeEsperaRepository;
 
 @Service
 public class SalaDeEsperaService {
 
-    @Autowired
-    private SalaDeEsperaRepository repository;
+    private final SalaDeEsperaRepository repository;
 
-    @Transactional
-    public SalaDeEspera insert(SalaDeEspera sala) {
-        SalaDeEspera novaSala;
-
-        try {
-            novaSala = repository.save(sala);
-        } catch (Exception e) {
-            throw new ObjectNotCreated(e);
-        }
-
-        return novaSala;
+    public SalaDeEsperaService(SalaDeEsperaRepository repository) {
+        this.repository = repository;
     }
 
     @Transactional
-    public SalaDeEspera findById(UUID id) {
-        Optional<SalaDeEspera> sala = repository.findById(id);
+    public SalaDeEspera insert(SalaDeEspera sala) {
+        try {
+            return repository.save(sala);
+        } catch (Exception e) {
+            throw new IllegalStateException("Erro ao criar Sala de Espera");
 
-        if (!sala.isPresent()) {
-            throw new NoSuchException("Sala de Espera");
         }
+    }
 
-        return sala.get();
+    @Transactional(readOnly = true)
+    public SalaDeEspera findById(UUID id) {
+        return repository.findById(id)
+                .orElseThrow(() -> new NoSuchException("Sala de Espera com ID: " + id));
+    }
+
+    @Transactional(readOnly = true)
+    public List<SalaDeEspera> findAll() {
+        return repository.findAll();
     }
 
     @Transactional(readOnly = true)
